move show_fixtures code to own module
